requirements.txt
python-dotenv>=0.15.0
sgqlc>=11.0
warrant>=0.6.1
